requirements.txt
Django==2.0.7
djangorestframework==3.8.2
fuzzywuzzy==0.16.0
python-Levenshtein==0.12.0
pytz==2018.5
